requirements.txt
Flask==2.3.3
flask-cors==6.0.2
groq==1.0.0
python-dotenv==1.0.0
gunicorn==23.0.0
